Cut chartOptionsBuilder to smaller pieces
-  Functions moved to new files according to the area they work in
-  axes, drilling, limits, series construction, tooltips

JIRA: RAIL-1815
