+++ libs/sdk-ui-charts/src/highcharts/_to_refactor/chartOptionsBuilder.ts
@@ -1,33 +1,15 @@
 // (C) 2007-2020 GoodData Corporation
-import { colors2Object, numberFormat } from "@gooddata/numberjs";
-import {
-    IAttributeDescriptor,
-    IDataView,
-    IMeasureDescriptor,
-    IMeasureGroupDescriptor,
-    IResultAttributeHeader,
-} from "@gooddata/sdk-backend-spi";
-import cx from "classnames";
+import { IDataView, IMeasureDescriptor, IMeasureGroupDescriptor } from "@gooddata/sdk-backend-spi";
 import invariant from "ts-invariant";
 
-import {
-    BucketNames,
-    DataViewFacade,
-    getDrillIntersection,
-    IHeaderPredicate,
-    IMappingHeader,
-    isSomeHeaderPredicateMatched,
-    VisType,
-    VisualizationTypes,
-} from "@gooddata/sdk-ui";
-import { IChartConfig, IChartLimits, ViewByAttributesLimit } from "../../interfaces";
+import { BucketNames, DataViewFacade, IHeaderPredicate } from "@gooddata/sdk-ui";
+import { IChartConfig, ViewByAttributesLimit } from "../../interfaces";
 import {
     PARENT_ATTRIBUTE_INDEX,
     PRIMARY_ATTRIBUTE_INDEX,
     STACK_BY_DIMENSION_INDEX,
     VIEW_BY_DIMENSION_INDEX,
 } from "./constants/dimensions";
-import { HEATMAP_DATA_POINTS_LIMIT, PIE_CHART_LIMIT } from "./constants/limits";
 
 import { findAttributeInDimension, findMeasureGroupInDimensions } from "./utils/executionResultHelper";
 import { IUnwrappedAttributeHeadersWithItems } from "./utils/types";
@@ -35,19 +17,14 @@ import { IUnwrappedAttributeHeadersWithItems } from "./utils/types";
 import { IColorStrategy } from "@gooddata/sdk-ui-vis-commons";
 
 import {
-    customEscape,
     isAreaChart,
-    isBarChart,
     isBubbleChart,
-    isBulletChart,
     isChartSupported,
     isComboChart,
-    isCssMultiLineTruncationSupported,
     isHeatmap,
     isOneOfTypes,
     isScatterPlot,
     isTreemap,
-    parseValue,
     stringifyChartTypes,
     unwrap,
 } from "./utils/common";
@@ -61,43 +38,30 @@ import {
 import { getCategoriesForTwoAttributes } from "./chartOptions/extendedStackingChartOptions";
 
 import { ColorFactory } from "./colorFactory";
-
-import {
-    DEFAULT_CATEGORIES_LIMIT,
-    DEFAULT_DATA_POINTS_LIMIT,
-    DEFAULT_SERIES_LIMIT,
-} from "../chartTypes/_integration/commonConfiguration";
 import { getChartProperties } from "../chartTypes/_integration/helpers";
 import Highcharts from "../chartTypes/_integration/highchartsEntryPoint";
-import { isDataOfReasonableSize } from "./highChartsCreators";
-import { formatValueForTooltip, getFormattedValueForTooltip } from "./tooltip";
-import { supportedDualAxesChartTypes } from "./chartCapabilities";
+import {
+    multiMeasuresAlternatingTypes,
+    sortedByMeasureTypes,
+    unsupportedStackingTypes,
+} from "./chartCapabilities";
 import cloneDeep from "lodash/cloneDeep";
-import compact from "lodash/compact";
 import get from "lodash/get";
-import includes from "lodash/includes";
 import isEmpty from "lodash/isEmpty";
 import isUndefined from "lodash/isUndefined";
-import range from "lodash/range";
 import without from "lodash/without";
-import omit from "lodash/omit";
 import { NORMAL_STACK, PERCENT_STACK } from "./constants/stacking";
+import { IChartOptions, IPointData } from "../typings/unsafe";
+import { getSeries } from "./chartSeries";
 import {
-    IAxis,
-    ICategory,
-    IChartOptions,
-    IPointData,
-    ISeriesDataItem,
-    ISeriesItem,
-    ISeriesItemConfig,
-} from "../typings/unsafe";
-import { getBulletChartSeries } from "../chartTypes/bulletChart/bulletChartSeries";
-import { getBubbleChartSeries } from "../chartTypes/bubbleChart/bubbleChartSeries";
-import { getTreemapStackedSeries } from "../chartTypes/treemap/treemapChartSeries";
-import { getHeatmapSeries } from "../chartTypes/heatmap/heatmapChartSeries";
-import { getScatterPlotSeries } from "../chartTypes/scatterPlot/scatterPlotSeries";
-
-const TOOLTIP_PADDING = 10;
+    buildTooltipFactory,
+    generateTooltipHeatmapFn,
+    generateTooltipXYFn,
+    getTooltipFactory,
+    ITooltipFactory,
+} from "./chartTooltips";
+import { getDrillableSeries } from "./chartDrilling";
+import { assignYAxes, getXAxes, getYAxes } from "./chartAxes";
 
 const isAreaChartStackingEnabled = (options: IChartConfig) => {
     const { type, stacking, stackMeasures } = options;
@@ -109,659 +73,6 @@ const isAreaChartStackingEnabled = (options: IChartConfig) => {
     }
     return stackMeasures;
 };
-
-// types with only many measures or one measure and one attribute
-const multiMeasuresAlternatingTypes = [
-    VisualizationTypes.PIE,
-    VisualizationTypes.DONUT,
-    VisualizationTypes.FUNNEL,
-    VisualizationTypes.TREEMAP,
-];
-
-const unsupportedNegativeValuesTypes = [
-    VisualizationTypes.PIE,
-    VisualizationTypes.DONUT,
-    VisualizationTypes.FUNNEL,
-    VisualizationTypes.TREEMAP,
-];
-
-// charts sorted by default by measure value
-const sortedByMeasureTypes = [VisualizationTypes.PIE, VisualizationTypes.DONUT, VisualizationTypes.FUNNEL];
-
-const unsupportedStackingTypes = [
-    VisualizationTypes.LINE,
-    VisualizationTypes.AREA,
-    VisualizationTypes.SCATTER,
-    VisualizationTypes.BUBBLE,
-];
-
-export interface IValidationResult {
-    dataTooLarge: boolean;
-    hasNegativeValue: boolean;
-}
-
-export type ITooltipFactory = (
-    point: IPointData,
-    maxTooltipContentWidth: number,
-    percentageValue?: number,
-) => string;
-
-export function isNegativeValueIncluded(series: ISeriesItem[]): boolean {
-    return series.some((seriesItem: ISeriesItem) =>
-        (seriesItem.data || []).some(({ y, value }: ISeriesDataItem) => y < 0 || value < 0),
-    );
-}
-
-function getChartLimits(type: string): IChartLimits {
-    switch (type) {
-        case VisualizationTypes.SCATTER:
-            return {
-                series: DEFAULT_SERIES_LIMIT,
-                categories: DEFAULT_SERIES_LIMIT,
-            };
-        case VisualizationTypes.PIE:
-        case VisualizationTypes.DONUT:
-        case VisualizationTypes.FUNNEL:
-            return {
-                series: 1,
-                categories: PIE_CHART_LIMIT,
-            };
-        case VisualizationTypes.TREEMAP:
-            return {
-                series: DEFAULT_SERIES_LIMIT,
-                categories: DEFAULT_DATA_POINTS_LIMIT,
-                dataPoints: DEFAULT_DATA_POINTS_LIMIT,
-            };
-        case VisualizationTypes.HEATMAP:
-            return {
-                series: DEFAULT_SERIES_LIMIT,
-                categories: DEFAULT_CATEGORIES_LIMIT,
-                dataPoints: HEATMAP_DATA_POINTS_LIMIT,
-            };
-        default:
-            return {
-                series: DEFAULT_SERIES_LIMIT,
-                categories: DEFAULT_CATEGORIES_LIMIT,
-            };
-    }
-}
-
-export function cannotShowNegativeValues(type: string): boolean {
-    return isOneOfTypes(type, unsupportedNegativeValuesTypes);
-}
-
-function getTreemapDataForValidation(data: any) {
-    // filter out root nodes
-    return {
-        ...data,
-        series: data.series.map((serie: any) => ({
-            ...serie,
-            data: serie.data.filter((dataItem: any) => dataItem.id === undefined),
-        })),
-    };
-}
-
-export function validateData(limits: IChartLimits, chartOptions: IChartOptions): IValidationResult {
-    const { type, isViewByTwoAttributes } = chartOptions;
-    const finalLimits = limits || getChartLimits(type);
-    let dataToValidate = chartOptions.data;
-    if (isTreemap(type)) {
-        dataToValidate = getTreemapDataForValidation(chartOptions.data);
-    }
-
-    return {
-        dataTooLarge: !isDataOfReasonableSize(dataToValidate, finalLimits, isViewByTwoAttributes),
-        hasNegativeValue: cannotShowNegativeValues(type) && isNegativeValueIncluded(chartOptions.data.series),
-    };
-}
-
-export function getSeriesItemData(
-    seriesItem: string[],
-    seriesIndex: number,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    type: string,
-    colorStrategy: IColorStrategy,
-): IPointData[] {
-    return seriesItem.map((pointValue: string, pointIndex: number) => {
-        // by default seriesIndex corresponds to measureGroup label index
-        let measureIndex = seriesIndex;
-        // by default pointIndex corresponds to viewBy label index
-        let viewByIndex = pointIndex;
-        // drillContext can have 1 to 3 items
-        // viewBy attribute label, stackby label if available
-        // last drillContextItem is always current serie measure
-        if (stackByAttribute) {
-            // pointIndex corresponds to viewBy attribute label (if available)
-            viewByIndex = pointIndex;
-            // stack bar chart has always just one measure
-            measureIndex = 0;
-        } else if (isOneOfTypes(type, multiMeasuresAlternatingTypes) && !viewByAttribute) {
-            measureIndex = pointIndex;
-        }
-
-        let valueProp: any = {
-            y: parseValue(pointValue),
-        };
-        if (isTreemap(type)) {
-            valueProp = {
-                value: parseValue(pointValue),
-            };
-        }
-
-        const pointData: IPointData = Object.assign(
-            {
-                ...valueProp,
-                format: unwrap(measureGroup.items[measureIndex]).format,
-            },
-            pointValue === null
-                ? {
-                      marker: {
-                          enabled: false,
-                      },
-                  }
-                : {},
-        );
-
-        if (stackByAttribute) {
-            // if there is a stackBy attribute, then seriesIndex corresponds to stackBy label index
-            pointData.name = unwrap(stackByAttribute.items[seriesIndex]).name;
-        } else if (isOneOfTypes(type, multiMeasuresAlternatingTypes) && viewByAttribute) {
-            pointData.name = unwrap(viewByAttribute.items[viewByIndex]).name;
-        } else {
-            pointData.name = unwrap(measureGroup.items[measureIndex]).name;
-        }
-
-        if (isOneOfTypes(type, multiMeasuresAlternatingTypes)) {
-            pointData.color = colorStrategy.getColorByIndex(pointIndex);
-            // Pie and Treemap charts use pointData viewByIndex as legendIndex if available
-            // instead of seriesItem legendIndex
-            pointData.legendIndex = viewByAttribute ? viewByIndex : pointIndex;
-        }
-
-        return pointData;
-    });
-}
-
-export function getSeries(
-    dv: DataViewFacade,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    type: string,
-    colorStrategy: IColorStrategy,
-): any {
-    if (isHeatmap(type)) {
-        return getHeatmapSeries(dv, measureGroup);
-    } else if (isScatterPlot(type)) {
-        return getScatterPlotSeries(dv, stackByAttribute, colorStrategy);
-    } else if (isBubbleChart(type)) {
-        return getBubbleChartSeries(dv, measureGroup, stackByAttribute, colorStrategy);
-    } else if (isTreemap(type) && stackByAttribute) {
-        return getTreemapStackedSeries(dv, measureGroup, viewByAttribute, stackByAttribute, colorStrategy);
-    } else if (isBulletChart(type)) {
-        return getBulletChartSeries(dv, measureGroup, colorStrategy);
-    }
-
-    return dv
-        .rawData()
-        .twoDimData()
-        .map((seriesItem: string[], seriesIndex: number) => {
-            const seriesItemData = getSeriesItemData(
-                seriesItem,
-                seriesIndex,
-                measureGroup,
-                viewByAttribute,
-                stackByAttribute,
-                type,
-                colorStrategy,
-            );
-
-            const seriesItemConfig: ISeriesItemConfig = {
-                color: colorStrategy.getColorByIndex(seriesIndex),
-                legendIndex: seriesIndex,
-                data: seriesItemData,
-            };
-
-            if (stackByAttribute) {
-                // if stackBy attribute is available, seriesName is a stackBy attribute value of index seriesIndex
-                // this is a limitiation of highcharts and a reason why you can not have multi-measure stacked charts
-                seriesItemConfig.name = stackByAttribute.items[seriesIndex].attributeHeaderItem.name;
-            } else if (isOneOfTypes(type, multiMeasuresAlternatingTypes) && !viewByAttribute) {
-                // Pie charts with measures only have a single series which name would is ambiguous
-                seriesItemConfig.name = measureGroup.items
-                    .map((wrappedMeasure: IMeasureDescriptor) => {
-                        return unwrap(wrappedMeasure).name;
-                    })
-                    .join(", ");
-            } else {
-                // otherwise seriesName is a measure name of index seriesIndex
-                seriesItemConfig.name = measureGroup.items[seriesIndex].measureHeaderItem.name;
-            }
-
-            const turboThresholdProp = isTreemap(type) ? { turboThreshold: 0 } : {};
-
-            return {
-                ...seriesItemConfig,
-                ...turboThresholdProp,
-            };
-        });
-}
-
-const renderTooltipHTML = (textData: string[][], maxTooltipContentWidth: number): string => {
-    const maxItemWidth = maxTooltipContentWidth - TOOLTIP_PADDING * 2;
-    const titleMaxWidth = maxItemWidth;
-    const multiLineTruncationSupported = isCssMultiLineTruncationSupported();
-    const threeDotsWidth = 16;
-    const valueMaxWidth = multiLineTruncationSupported ? maxItemWidth : maxItemWidth - threeDotsWidth;
-    const titleStyle = `style="max-width: ${titleMaxWidth}px;"`;
-    const valueStyle = `style="max-width: ${valueMaxWidth}px;"`;
-    const itemClass = cx("gd-viz-tooltip-item", {
-        "multiline-supported": multiLineTruncationSupported,
-    });
-    const valueClass = cx("gd-viz-tooltip-value", {
-        "clamp-two-line": multiLineTruncationSupported,
-    });
-
-    return textData
-        .map((item: string[]) => {
-            // the third span is hidden, that help to have tooltip work with max-width
-            return `<div class="${itemClass}">
-                        <span class="gd-viz-tooltip-title" ${titleStyle}>${item[0]}</span>
-                        <div class="gd-viz-tooltip-value-wraper" ${titleStyle}>
-                            <span class="${valueClass}" ${valueStyle}>${item[1]}</span>
-                        </div>
-                        <div class="gd-viz-tooltip-value-wraper" ${titleStyle}>
-                            <span class="gd-viz-tooltip-value-max-content" ${valueStyle}>${item[1]}</span>
-                        </div>
-                    </div>`;
-        })
-        .join("\n");
-};
-
-function isPointOnOppositeAxis(point: IPointData): boolean {
-    return get(point, ["series", "yAxis", "opposite"], false);
-}
-
-export function buildTooltipFactory(
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    type: string,
-    config: IChartConfig = {},
-    isDualAxis: boolean = false,
-): ITooltipFactory {
-    const { separators, stackMeasuresToPercent = false } = config;
-
-    return (point: IPointData, maxTooltipContentWidth: number, percentageValue?: number): string => {
-        const isDualChartWithRightAxis = isDualAxis && isPointOnOppositeAxis(point);
-        const formattedValue = getFormattedValueForTooltip(
-            isDualChartWithRightAxis,
-            stackMeasuresToPercent,
-            point,
-            separators,
-            percentageValue,
-        );
-
-        const textData = [[customEscape(point.series.name), formattedValue]];
-
-        if (viewByAttribute) {
-            // For some reason, highcharts ommit categories for pie charts with attribute. Use point.name instead.
-            // use attribute name instead of attribute display form name
-            textData.unshift([
-                customEscape(viewByAttribute.formOf.name),
-                // since applying 'grouped-categories' plugin,
-                // 'category' type is replaced from string to object in highchart
-                customEscape((point.category && point.category.name) || point.name),
-            ]);
-        } else if (isOneOfTypes(type, multiMeasuresAlternatingTypes)) {
-            // Pie charts with measure only have to use point.name instead of series.name to get the measure name
-            textData[0][0] = customEscape(point.name);
-        }
-        return renderTooltipHTML(textData, maxTooltipContentWidth);
-    };
-}
-
-export function buildTooltipForTwoAttributesFactory(
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    viewByParentAttribute: IUnwrappedAttributeHeadersWithItems,
-    config: IChartConfig = {},
-    isDualAxis: boolean = false,
-): ITooltipFactory {
-    const { separators, stackMeasuresToPercent = false } = config;
-
-    return (point: IPointData, maxTooltipContentWidth: number, percentageValue?: number): string => {
-        const category: ICategory = point.category;
-
-        const isDualChartWithRightAxis = isDualAxis && isPointOnOppositeAxis(point);
-        const formattedValue = getFormattedValueForTooltip(
-            isDualChartWithRightAxis,
-            stackMeasuresToPercent,
-            point,
-            separators,
-            percentageValue,
-        );
-
-        const textData = [[customEscape(point.series.name), formattedValue]];
-
-        if (category) {
-            if (viewByAttribute) {
-                textData.unshift([customEscape(viewByAttribute.formOf.name), customEscape(category.name)]);
-            }
-
-            if (viewByParentAttribute && category.parent) {
-                textData.unshift([
-                    customEscape(viewByParentAttribute.formOf.name),
-                    customEscape(category.parent.name),
-                ]);
-            }
-        }
-
-        return renderTooltipHTML(textData, maxTooltipContentWidth);
-    };
-}
-
-export function generateTooltipXYFn(
-    measures: IMeasureDescriptor[],
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    config: IChartConfig = {},
-): ITooltipFactory {
-    const { separators } = config;
-
-    return (point: IPointData, maxTooltipContentWidth: number): string => {
-        const textData = [];
-        const name = point.name ? point.name : point.series.name;
-
-        if (stackByAttribute) {
-            textData.unshift([customEscape(stackByAttribute.formOf.name), customEscape(name)]);
-        }
-
-        if (measures[0]) {
-            textData.push([
-                customEscape(measures[0].measureHeaderItem.name),
-                formatValueForTooltip(point.x, measures[0].measureHeaderItem.format, separators),
-            ]);
-        }
-
-        if (measures[1]) {
-            textData.push([
-                customEscape(measures[1].measureHeaderItem.name),
-                formatValueForTooltip(point.y, measures[1].measureHeaderItem.format, separators),
-            ]);
-        }
-
-        if (measures[2]) {
-            textData.push([
-                customEscape(measures[2].measureHeaderItem.name),
-                formatValueForTooltip(point.z, measures[2].measureHeaderItem.format, separators),
-            ]);
-        }
-
-        return renderTooltipHTML(textData, maxTooltipContentWidth);
-    };
-}
-
-export function generateTooltipHeatmapFn(
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    viewByAttribute: any,
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    stackByAttribute: any,
-    config: IChartConfig = {},
-): ITooltipFactory {
-    const { separators } = config;
-    const formatValue = (val: number, format: string) => {
-        return colors2Object(val === null ? "-" : numberFormat(val, format, undefined, separators));
-    };
-
-    return (point: IPointData, maxTooltipContentWidth: number): string => {
-        const formattedValue = customEscape(
-            formatValue(point.value, point.series.userOptions.dataLabels.formatGD).label,
-        );
-        const textData = [];
-
-        textData.unshift([customEscape(point.series.name), formattedValue]);
-
-        if (viewByAttribute) {
-            textData.unshift([
-                customEscape(viewByAttribute.formOf.name),
-                customEscape(viewByAttribute.items[point.x].attributeHeaderItem.name),
-            ]);
-        }
-        if (stackByAttribute) {
-            textData.unshift([
-                customEscape(stackByAttribute.formOf.name),
-                customEscape(stackByAttribute.items[point.y].attributeHeaderItem.name),
-            ]);
-        }
-
-        return renderTooltipHTML(textData, maxTooltipContentWidth);
-    };
-}
-
-export function buildTooltipTreemapFactory(
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    config: IChartConfig = {},
-): ITooltipFactory {
-    const { separators } = config;
-
-    return (point: IPointData, maxTooltipContentWidth: number) => {
-        // show tooltip for leaf node only
-        if (!point.node || point.node.isLeaf === false) {
-            return null;
-        }
-        const formattedValue = formatValueForTooltip(point.value, point.format, separators);
-
-        const textData = [];
-
-        if (stackByAttribute) {
-            textData.push([
-                customEscape(stackByAttribute.formOf.name),
-                customEscape(stackByAttribute.items[point.y].attributeHeaderItem.name),
-            ]);
-        }
-
-        if (viewByAttribute) {
-            textData.unshift([
-                customEscape(viewByAttribute.formOf.name),
-                customEscape(viewByAttribute.items[point.x].attributeHeaderItem.name),
-            ]);
-            textData.push([customEscape(point.series.name), formattedValue]);
-        } else {
-            textData.push([customEscape(point.category && point.category.name), formattedValue]);
-        }
-
-        return renderTooltipHTML(textData, maxTooltipContentWidth);
-    };
-}
-
-export interface ILegacyMeasureHeader {
-    uri: string; // header attribute value or measure uri
-    identifier?: string;
-    localIdentifier: string;
-    name: string; // header attribute value or measure text label
-}
-
-export interface ILegacyAttributeHeader extends ILegacyMeasureHeader {
-    attribute: any;
-}
-
-export type ILegacyHeader = ILegacyAttributeHeader | ILegacyMeasureHeader;
-
-export function isLegacyAttributeHeader(header: ILegacyHeader): header is ILegacyAttributeHeader {
-    return (header as ILegacyAttributeHeader).attribute !== undefined;
-}
-
-function getViewBy(viewByAttribute: IUnwrappedAttributeHeadersWithItems, viewByIndex: number) {
-    let viewByHeader: IResultAttributeHeader = null;
-    let viewByItem = null;
-    let viewByAttributeDescriptor: IAttributeDescriptor = null;
-
-    if (viewByAttribute) {
-        viewByHeader = viewByAttribute.items[viewByIndex];
-        viewByItem = {
-            ...unwrap(viewByHeader),
-            attribute: viewByAttribute,
-        };
-        viewByAttributeDescriptor = { attributeHeader: omit(viewByAttribute, "items") };
-    }
-
-    return {
-        viewByHeader,
-        viewByItem,
-        viewByAttributeDescriptor,
-    };
-}
-
-function getStackBy(stackByAttribute: IUnwrappedAttributeHeadersWithItems, stackByIndex: number) {
-    let stackByHeader: IResultAttributeHeader = null;
-    let stackByItem = null;
-    let stackByAttributeDescriptor: IAttributeDescriptor = null;
-
-    if (stackByAttribute) {
-        // stackBy item index is always equal to seriesIndex
-        stackByHeader = stackByAttribute.items[stackByIndex];
-        stackByItem = {
-            ...unwrap(stackByHeader),
-            attribute: stackByAttribute,
-        };
-        stackByAttributeDescriptor = { attributeHeader: omit(stackByAttribute, "items") };
-    }
-
-    return {
-        stackByHeader,
-        stackByItem,
-        stackByAttributeDescriptor,
-    };
-}
-
-export function getDrillableSeries(
-    dv: DataViewFacade,
-    series: any[],
-    drillableItems: IHeaderPredicate[],
-    viewByAttributes: IUnwrappedAttributeHeadersWithItems[],
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    type: VisType,
-): any {
-    const [viewByChildAttribute, viewByParentAttribute] = viewByAttributes;
-
-    const isMultiMeasureWithOnlyMeasures =
-        isOneOfTypes(type, multiMeasuresAlternatingTypes) && !viewByChildAttribute;
-    const measureGroup = findMeasureGroupInDimensions(dv.meta().dimensions());
-
-    return series.map((seriesItem, seriesIndex) => {
-        let isSeriesDrillable = false;
-        let data =
-            seriesItem.data &&
-            seriesItem.data.map((pointData: IPointData, pointIndex: number) => {
-                let measureHeaders: IMappingHeader[] = [];
-
-                const isStackedTreemap = isTreemap(type) && !!stackByAttribute;
-                if (isScatterPlot(type)) {
-                    measureHeaders = get(measureGroup, "items", []).slice(0, 2);
-                } else if (isBubbleChart(type)) {
-                    measureHeaders = get(measureGroup, "items", []).slice(0, 3);
-                } else if (isStackedTreemap) {
-                    if (pointData.id !== undefined) {
-                        // not leaf -> can't be drillable
-                        return pointData;
-                    }
-                    const measureIndex = viewByChildAttribute ? 0 : parseInt(pointData.parent, 10);
-                    measureHeaders = [measureGroup.items[measureIndex]];
-                } else {
-                    // measureIndex is usually seriesIndex,
-                    // except for stack by attribute and metricOnly pie or donut chart
-                    // it is looped-around pointIndex instead
-                    // Looping around the end of items array only works when
-                    // measureGroup is the last header on it's dimension
-                    // We do not support setups with measureGroup before attributeHeaders
-                    const measureIndex =
-                        !stackByAttribute && !isMultiMeasureWithOnlyMeasures
-                            ? seriesIndex
-                            : pointIndex % measureGroup.items.length;
-                    measureHeaders = [measureGroup.items[measureIndex]];
-                }
-
-                const viewByIndex = isHeatmap(type) || isStackedTreemap ? pointData.x : pointIndex;
-                let stackByIndex = isHeatmap(type) || isStackedTreemap ? pointData.y : seriesIndex;
-                if (isScatterPlot(type)) {
-                    stackByIndex = viewByIndex; // scatter plot uses stack by attribute but has only one serie
-                }
-
-                const { stackByHeader, stackByAttributeDescriptor } = getStackBy(
-                    stackByAttribute,
-                    stackByIndex,
-                );
-
-                const {
-                    viewByHeader: viewByChildHeader,
-                    viewByAttributeDescriptor: viewByChildAttributeDescriptor,
-                } = getViewBy(viewByChildAttribute, viewByIndex);
-
-                const {
-                    viewByHeader: viewByParentHeader,
-                    viewByAttributeDescriptor: viewByParentdAttributeDescriptor,
-                } = getViewBy(viewByParentAttribute, viewByIndex);
-
-                // point is drillable if a drillableItem matches:
-                //   point's measure,
-                //   point's viewBy attribute,
-                //   point's viewBy attribute item,
-                //   point's stackBy attribute,
-                //   point's stackBy attribute item,
-                const drillableHooks: IMappingHeader[] = without(
-                    [
-                        ...measureHeaders,
-                        viewByChildAttributeDescriptor,
-                        viewByChildHeader,
-                        viewByParentdAttributeDescriptor,
-                        viewByParentHeader,
-                        stackByAttributeDescriptor,
-                        stackByHeader,
-                    ],
-                    null,
-                );
-
-                const drilldown: boolean = drillableHooks.some((drillableHook) =>
-                    isSomeHeaderPredicateMatched(drillableItems, drillableHook, dv),
-                );
-
-                const drillableProps: any = {
-                    drilldown,
-                };
-
-                if (drilldown) {
-                    const headers: IMappingHeader[] = [
-                        ...measureHeaders,
-                        viewByChildHeader,
-                        viewByChildAttributeDescriptor,
-                        viewByParentHeader,
-                        viewByParentdAttributeDescriptor,
-                        stackByHeader,
-                        stackByAttributeDescriptor,
-                    ];
-                    const sanitizedHeaders = without([...headers], null);
-                    drillableProps.drillIntersection = getDrillIntersection(sanitizedHeaders);
-                    isSeriesDrillable = true;
-                }
-                return {
-                    ...pointData,
-                    ...drillableProps,
-                };
-            });
-
-        if (isScatterPlot(type)) {
-            data = data.filter((dataItem: ISeriesDataItem) => {
-                return dataItem.x !== null && dataItem.y !== null;
-            });
-        }
-
-        return {
-            ...seriesItem,
-            data,
-            isDrillable: isSeriesDrillable,
-        };
-    });
-}
 
 function getCategories(
     type: string,
@@ -819,244 +130,6 @@ function getStackingConfig(
         return stackingValue;
     }
     return null; // no stacking
-}
-
-function preprocessMeasureGroupItems(
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    defaultValues: any,
-): any[] {
-    return measureGroup.items.map((item: IMeasureDescriptor, index: number) => {
-        const unwrapped = unwrap(item);
-        return index
-            ? {
-                  label: unwrapped.name,
-                  format: unwrapped.format,
-              }
-            : {
-                  label: defaultValues.label || unwrapped.name,
-                  format: defaultValues.format || unwrapped.format,
-              };
-    });
-}
-
-function getXAxes(
-    dv: DataViewFacade,
-    config: IChartConfig,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-): IAxis[] {
-    const { type } = config;
-
-    if (isScatterPlot(type) || isBubbleChart(type)) {
-        const measureGroupItems = preprocessMeasureGroupItems(measureGroup, {
-            label: config.xLabel,
-            format: config.xFormat,
-        });
-
-        const firstMeasureGroupItem = measureGroupItems[0];
-
-        const noPrimaryMeasures = dv.def().isBucketEmpty(BucketNames.MEASURES);
-        if (noPrimaryMeasures) {
-            return [
-                {
-                    label: "",
-                },
-            ];
-        } else {
-            return [
-                {
-                    label: firstMeasureGroupItem.label || "",
-                    format: firstMeasureGroupItem.format || "",
-                },
-            ];
-        }
-    }
-
-    const xLabel = config.xLabel || (viewByAttribute ? viewByAttribute.formOf.name : "");
-    return [
-        {
-            label: xLabel,
-        },
-    ];
-}
-
-function getMeasureFormatKey(measureGroupItems: IMeasureDescriptor[]) {
-    const percentageFormat = getMeasureFormat(
-        measureGroupItems.find((measure: any) => {
-            return isPercentage(getMeasureFormat(measure));
-        }),
-    );
-    return percentageFormat !== ""
-        ? {
-              format: percentageFormat,
-          }
-        : {};
-}
-
-function getMeasureFormat(measure: any) {
-    return get(measure, "format", "");
-}
-
-function isPercentage(format: string) {
-    return format.includes("%");
-}
-
-function getYAxes(
-    dv: DataViewFacade,
-    config: IChartConfig,
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-    stackByAttribute: any,
-): IAxis[] {
-    const { type } = config;
-
-    const measureGroupItems = preprocessMeasureGroupItems(measureGroup, {
-        label: config.yLabel,
-        format: config.yFormat,
-    });
-
-    const firstMeasureGroupItem = measureGroupItems[0];
-    const secondMeasureGroupItem = measureGroupItems[1];
-    const hasMoreThanOneMeasure = measureGroupItems.length > 1;
-    const noPrimaryMeasures = dv.def().isBucketEmpty(BucketNames.MEASURES);
-
-    const { measures: secondaryAxisMeasures = [] as string[] } =
-        (isBarChart(type) ? config.secondary_xaxis : config.secondary_yaxis) || {};
-
-    let yAxes: IAxis[] = [];
-
-    if (isScatterPlot(type) || isBubbleChart(type)) {
-        const hasSecondaryMeasure = !dv.def().isBucketEmpty(BucketNames.SECONDARY_MEASURES);
-
-        if (hasSecondaryMeasure) {
-            if (noPrimaryMeasures) {
-                yAxes = [
-                    {
-                        ...firstMeasureGroupItem,
-                    },
-                ];
-            } else {
-                yAxes = [
-                    {
-                        ...secondMeasureGroupItem,
-                    },
-                ];
-            }
-        } else {
-            yAxes = [{ label: "" }];
-        }
-    } else if (isHeatmap(type)) {
-        yAxes = [
-            {
-                label: stackByAttribute ? stackByAttribute.formOf.name : "",
-            },
-        ];
-    } else if (
-        isOneOfTypes(type, supportedDualAxesChartTypes) &&
-        !isEmpty(measureGroupItems) &&
-        !isEmpty(secondaryAxisMeasures)
-    ) {
-        const { measuresInFirstAxis, measuresInSecondAxis }: IMeasuresInAxes = assignMeasuresToAxes(
-            secondaryAxisMeasures,
-            measureGroup,
-        );
-
-        let firstAxis: IAxis = createYAxisItem(measuresInFirstAxis, false);
-        let secondAxis: IAxis = createYAxisItem(measuresInSecondAxis, true);
-
-        if (firstAxis) {
-            firstAxis = {
-                ...firstAxis,
-                ...getMeasureFormatKey(measuresInFirstAxis),
-                seriesIndices: measuresInFirstAxis.map(({ index }: any) => index),
-            };
-        }
-        if (secondAxis) {
-            secondAxis = {
-                ...secondAxis,
-                ...getMeasureFormatKey(measuresInSecondAxis),
-                seriesIndices: measuresInSecondAxis.map(({ index }: any) => index),
-            };
-        }
-
-        yAxes = compact([firstAxis, secondAxis]);
-    } else {
-        // if more than one measure and NOT dual, then have empty item name
-        const nonDualMeasureAxis = hasMoreThanOneMeasure
-            ? {
-                  label: "",
-              }
-            : {};
-        yAxes = [
-            {
-                ...firstMeasureGroupItem,
-                ...nonDualMeasureAxis,
-                seriesIndices: range(measureGroupItems.length),
-                ...getMeasureFormatKey(measureGroupItems),
-            },
-        ];
-    }
-
-    return yAxes;
-}
-
-interface IMeasuresInAxes {
-    measuresInFirstAxis: any[];
-    measuresInSecondAxis: any[];
-}
-
-function assignMeasuresToAxes(
-    secondMeasures: string[],
-    measureGroup: IMeasureGroupDescriptor["measureGroupHeader"],
-): IMeasuresInAxes {
-    return measureGroup.items.reduce(
-        (
-            result: any,
-            { measureHeaderItem: { name, format, localIdentifier } }: IMeasureDescriptor,
-            index,
-        ) => {
-            if (includes(secondMeasures, localIdentifier)) {
-                result.measuresInSecondAxis.push({ name, format, index });
-            } else {
-                result.measuresInFirstAxis.push({ name, format, index });
-            }
-            return result;
-        },
-        {
-            measuresInFirstAxis: [],
-            measuresInSecondAxis: [],
-        },
-    );
-}
-
-function createYAxisItem(measuresInAxis: any[], opposite = false) {
-    const length = measuresInAxis.length;
-    if (length) {
-        const { name, format } = measuresInAxis[0];
-        return {
-            label: length === 1 ? name : "",
-            format,
-            opposite,
-        };
-    }
-    return null;
-}
-
-function assignYAxes(series: any, yAxes: IAxis[]) {
-    return series.reduce((result: any, item: any, index: number) => {
-        const yAxisIndex = yAxes.findIndex((axis: IAxis) => {
-            return includes(get(axis, "seriesIndices", []), index);
-        });
-        // for case viewBy and stackBy have one attribute, and one measure is sliced to multiple series
-        // then 'yAxis' in other series should follow the first one
-        const firstYAxisIndex = result.lenght > 0 ? result[0].yAxis : 0;
-        const seriesItem = {
-            ...item,
-            yAxis: yAxisIndex !== -1 ? yAxisIndex : firstYAxisIndex,
-        };
-
-        result.push(seriesItem);
-        return result;
-    }, []);
 }
 
 export const HEAT_MAP_CATEGORIES_COUNT = 7;
@@ -1175,29 +248,6 @@ export function getTreemapAttributes(dv: DataViewFacade): any {
             1,
         ),
     };
-}
-
-function getTooltipFactory(
-    isViewByTwoAttributes: boolean,
-    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
-    viewByParentAttribute: IUnwrappedAttributeHeadersWithItems,
-    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
-    config: IChartConfig = {},
-    isDualAxis: boolean = false,
-): ITooltipFactory {
-    const { type } = config;
-    if (isTreemap(type)) {
-        return buildTooltipTreemapFactory(viewByAttribute, stackByAttribute, config);
-    }
-    if (isViewByTwoAttributes) {
-        return buildTooltipForTwoAttributesFactory(
-            viewByAttribute,
-            viewByParentAttribute,
-            config,
-            isDualAxis,
-        );
-    }
-    return buildTooltipFactory(viewByAttribute, type, config, isDualAxis);
 }
 
 export function getChartOptions(
