+++ libs/sdk-ui-charts/src/highcharts/_to_refactor/chartTooltips.ts
@@ -1,0 +1,265 @@
+// (C) 2007-2020 GoodData Corporation
+import { IChartConfig } from "../../interfaces";
+import { colors2Object, numberFormat } from "@gooddata/numberjs";
+import { ICategory, IPointData } from "../typings/unsafe";
+import { customEscape, isCssMultiLineTruncationSupported, isOneOfTypes, isTreemap } from "./utils/common";
+import { IUnwrappedAttributeHeadersWithItems } from "./utils/types";
+import { formatValueForTooltip, getFormattedValueForTooltip } from "./tooltip";
+import { multiMeasuresAlternatingTypes } from "./chartCapabilities";
+import cx from "classnames";
+import get from "lodash/get";
+import { IMeasureDescriptor } from "@gooddata/sdk-backend-spi";
+
+const TOOLTIP_PADDING = 10;
+export type ITooltipFactory = (
+    point: IPointData,
+    maxTooltipContentWidth: number,
+    percentageValue?: number,
+) => string;
+const renderTooltipHTML = (textData: string[][], maxTooltipContentWidth: number): string => {
+    const maxItemWidth = maxTooltipContentWidth - TOOLTIP_PADDING * 2;
+    const titleMaxWidth = maxItemWidth;
+    const multiLineTruncationSupported = isCssMultiLineTruncationSupported();
+    const threeDotsWidth = 16;
+    const valueMaxWidth = multiLineTruncationSupported ? maxItemWidth : maxItemWidth - threeDotsWidth;
+    const titleStyle = `style="max-width: ${titleMaxWidth}px;"`;
+    const valueStyle = `style="max-width: ${valueMaxWidth}px;"`;
+    const itemClass = cx("gd-viz-tooltip-item", {
+        "multiline-supported": multiLineTruncationSupported,
+    });
+    const valueClass = cx("gd-viz-tooltip-value", {
+        "clamp-two-line": multiLineTruncationSupported,
+    });
+
+    return textData
+        .map((item: string[]) => {
+            // the third span is hidden, that help to have tooltip work with max-width
+            return `<div class="${itemClass}">
+                        <span class="gd-viz-tooltip-title" ${titleStyle}>${item[0]}</span>
+                        <div class="gd-viz-tooltip-value-wraper" ${titleStyle}>
+                            <span class="${valueClass}" ${valueStyle}>${item[1]}</span>
+                        </div>
+                        <div class="gd-viz-tooltip-value-wraper" ${titleStyle}>
+                            <span class="gd-viz-tooltip-value-max-content" ${valueStyle}>${item[1]}</span>
+                        </div>
+                    </div>`;
+        })
+        .join("\n");
+};
+
+function isPointOnOppositeAxis(point: IPointData): boolean {
+    return get(point, ["series", "yAxis", "opposite"], false);
+}
+
+export function buildTooltipFactory(
+    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
+    type: string,
+    config: IChartConfig = {},
+    isDualAxis: boolean = false,
+): ITooltipFactory {
+    const { separators, stackMeasuresToPercent = false } = config;
+
+    return (point: IPointData, maxTooltipContentWidth: number, percentageValue?: number): string => {
+        const isDualChartWithRightAxis = isDualAxis && isPointOnOppositeAxis(point);
+        const formattedValue = getFormattedValueForTooltip(
+            isDualChartWithRightAxis,
+            stackMeasuresToPercent,
+            point,
+            separators,
+            percentageValue,
+        );
+
+        const textData = [[customEscape(point.series.name), formattedValue]];
+
+        if (viewByAttribute) {
+            // For some reason, highcharts ommit categories for pie charts with attribute. Use point.name instead.
+            // use attribute name instead of attribute display form name
+            textData.unshift([
+                customEscape(viewByAttribute.formOf.name),
+                // since applying 'grouped-categories' plugin,
+                // 'category' type is replaced from string to object in highchart
+                customEscape((point.category && point.category.name) || point.name),
+            ]);
+        } else if (isOneOfTypes(type, multiMeasuresAlternatingTypes)) {
+            // Pie charts with measure only have to use point.name instead of series.name to get the measure name
+            textData[0][0] = customEscape(point.name);
+        }
+        return renderTooltipHTML(textData, maxTooltipContentWidth);
+    };
+}
+
+export function buildTooltipForTwoAttributesFactory(
+    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
+    viewByParentAttribute: IUnwrappedAttributeHeadersWithItems,
+    config: IChartConfig = {},
+    isDualAxis: boolean = false,
+): ITooltipFactory {
+    const { separators, stackMeasuresToPercent = false } = config;
+
+    return (point: IPointData, maxTooltipContentWidth: number, percentageValue?: number): string => {
+        const category: ICategory = point.category;
+
+        const isDualChartWithRightAxis = isDualAxis && isPointOnOppositeAxis(point);
+        const formattedValue = getFormattedValueForTooltip(
+            isDualChartWithRightAxis,
+            stackMeasuresToPercent,
+            point,
+            separators,
+            percentageValue,
+        );
+
+        const textData = [[customEscape(point.series.name), formattedValue]];
+
+        if (category) {
+            if (viewByAttribute) {
+                textData.unshift([customEscape(viewByAttribute.formOf.name), customEscape(category.name)]);
+            }
+
+            if (viewByParentAttribute && category.parent) {
+                textData.unshift([
+                    customEscape(viewByParentAttribute.formOf.name),
+                    customEscape(category.parent.name),
+                ]);
+            }
+        }
+
+        return renderTooltipHTML(textData, maxTooltipContentWidth);
+    };
+}
+
+export function generateTooltipXYFn(
+    measures: IMeasureDescriptor[],
+    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
+    config: IChartConfig = {},
+): ITooltipFactory {
+    const { separators } = config;
+
+    return (point: IPointData, maxTooltipContentWidth: number): string => {
+        const textData = [];
+        const name = point.name ? point.name : point.series.name;
+
+        if (stackByAttribute) {
+            textData.unshift([customEscape(stackByAttribute.formOf.name), customEscape(name)]);
+        }
+
+        if (measures[0]) {
+            textData.push([
+                customEscape(measures[0].measureHeaderItem.name),
+                formatValueForTooltip(point.x, measures[0].measureHeaderItem.format, separators),
+            ]);
+        }
+
+        if (measures[1]) {
+            textData.push([
+                customEscape(measures[1].measureHeaderItem.name),
+                formatValueForTooltip(point.y, measures[1].measureHeaderItem.format, separators),
+            ]);
+        }
+
+        if (measures[2]) {
+            textData.push([
+                customEscape(measures[2].measureHeaderItem.name),
+                formatValueForTooltip(point.z, measures[2].measureHeaderItem.format, separators),
+            ]);
+        }
+
+        return renderTooltipHTML(textData, maxTooltipContentWidth);
+    };
+}
+
+export function generateTooltipHeatmapFn(
+    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+    viewByAttribute: any,
+    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+    stackByAttribute: any,
+    config: IChartConfig = {},
+): ITooltipFactory {
+    const { separators } = config;
+    const formatValue = (val: number, format: string) => {
+        return colors2Object(val === null ? "-" : numberFormat(val, format, undefined, separators));
+    };
+
+    return (point: IPointData, maxTooltipContentWidth: number): string => {
+        const formattedValue = customEscape(
+            formatValue(point.value, point.series.userOptions.dataLabels.formatGD).label,
+        );
+        const textData = [];
+
+        textData.unshift([customEscape(point.series.name), formattedValue]);
+
+        if (viewByAttribute) {
+            textData.unshift([
+                customEscape(viewByAttribute.formOf.name),
+                customEscape(viewByAttribute.items[point.x].attributeHeaderItem.name),
+            ]);
+        }
+        if (stackByAttribute) {
+            textData.unshift([
+                customEscape(stackByAttribute.formOf.name),
+                customEscape(stackByAttribute.items[point.y].attributeHeaderItem.name),
+            ]);
+        }
+
+        return renderTooltipHTML(textData, maxTooltipContentWidth);
+    };
+}
+
+export function buildTooltipTreemapFactory(
+    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
+    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
+    config: IChartConfig = {},
+): ITooltipFactory {
+    const { separators } = config;
+
+    return (point: IPointData, maxTooltipContentWidth: number) => {
+        // show tooltip for leaf node only
+        if (!point.node || point.node.isLeaf === false) {
+            return null;
+        }
+        const formattedValue = formatValueForTooltip(point.value, point.format, separators);
+
+        const textData = [];
+
+        if (stackByAttribute) {
+            textData.push([
+                customEscape(stackByAttribute.formOf.name),
+                customEscape(stackByAttribute.items[point.y].attributeHeaderItem.name),
+            ]);
+        }
+
+        if (viewByAttribute) {
+            textData.unshift([
+                customEscape(viewByAttribute.formOf.name),
+                customEscape(viewByAttribute.items[point.x].attributeHeaderItem.name),
+            ]);
+            textData.push([customEscape(point.series.name), formattedValue]);
+        } else {
+            textData.push([customEscape(point.category && point.category.name), formattedValue]);
+        }
+
+        return renderTooltipHTML(textData, maxTooltipContentWidth);
+    };
+}
+
+export function getTooltipFactory(
+    isViewByTwoAttributes: boolean,
+    viewByAttribute: IUnwrappedAttributeHeadersWithItems,
+    viewByParentAttribute: IUnwrappedAttributeHeadersWithItems,
+    stackByAttribute: IUnwrappedAttributeHeadersWithItems,
+    config: IChartConfig = {},
+    isDualAxis: boolean = false,
+): ITooltipFactory {
+    const { type } = config;
+    if (isTreemap(type)) {
+        return buildTooltipTreemapFactory(viewByAttribute, stackByAttribute, config);
+    }
+    if (isViewByTwoAttributes) {
+        return buildTooltipForTwoAttributesFactory(
+            viewByAttribute,
+            viewByParentAttribute,
+            config,
+            isDualAxis,
+        );
+    }
+    return buildTooltipFactory(viewByAttribute, type, config, isDualAxis);
+}
